App Manager edits
Much more powerful applications tab. I've added the ability to pull the real services and relationships from the names in the listbox, but I need someone with expertise on running them with the Pis to finish the Evaluate(Command command) function. The entire thing should only be a couple lines.